requirement.txt
fastapi
uvicorn
sqlalchemy
pymongo
mysql-connector-python
psycopg2-binary
pyodbc
langchain
langchain-openai
langchain-community
langchain-groq
faiss-cpu
passlib
python-jose[cryptography]
bcrypt==3.2.0
pyjwt
cryptography
soundfile
sounddevice
numpy
groq
python-dotenv
pydantic
python-multipart